requirements.txt
# Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# PDF Parsing
docling>=2.0.0

# Embeddings & LLM
openai>=1.30.0

# Vector Store (with built-in BM25/sparse via fastembed)
qdrant-client[fastembed]>=1.9.0

# Backend
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic-settings>=2.2.0

# Frontend
streamlit>=1.35.0

# Config
pyyaml>=6.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0